+++ src/lib.rs
@@ -38,6 +38,7 @@ mod frameworks;
 mod fs;
 mod gdb;
 mod gles;
+mod host_cpu;
 mod image;
 mod libc;
 mod licenses;